requirements.txt
yfinance
requests
beautifulsoup4 
pandas
psycopg2-binary
sqlalchemy
dbt-postgres
dbt-core
dbt-clickhouse
apache-airflow-providers-airbyte==3.3.2
# airbyte-api